fix: media metadata error refactor: extract store to a file
