+++ electron_src/main.js
@@ -2,25 +2,13 @@ const electron = require('electron');
 // import reloader from 'electron-reloader';
 const { fixCORS } = require('./cors');
 const isDev = require('./isDev');
-const Store = require('electron-store');
+const store = require('./store');
 const { app, BrowserWindow, ipcMain, session } = electron;
 // isDev && reloader(module);
 if (isDev) {
   require('electron-reloader')(module);
 }
-const schema = {
-  theme: { type: 'string', default: 'black' },
-  windowState: {
-    type: 'object',
-    default: {
-      width: 1000,
-      height: 670,
-      maximized: false,
-      zoomLevel: 0
-    }
-  }
-};
-const store = new Store({ schema });
+
 const theme = store.get('theme');
 /** @type {{ width: number; height: number; maximized: boolean; zoomLevel: number}} */
 const windowState = store.get('windowState');
