+++ electron_src/package.json
@@ -91,7 +91,7 @@
     }
   },
   "devDependencies": {
-    "electron": "14.2.0",
+    "electron": "14.2.1",
     "electron-builder": "^22.13.1",
     "electron-reloader": "^1.2.1"
   },
